pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cognisim"
version = "0.1.0"
description = "A package for mobile app adaptation and testing"
authors = ["Revyl AI <anam@revyl.ai>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/RevylAI/Mobileadapt"
homepage = "https://mobileadapt.revyl.ai"
packages = [{include = "cognisim"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]
python = "^3.12"
appium-python-client = "*"
loguru = "*"
lxml = "*"
numpy = "*"
attrs = "*"
str2bool = "^1.1"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
retrying = "^1.3.4"
openai = "^1.43.0"
setuptools = "^75.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/RevylAI/Mobileadapt/issues"
[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"

